moving files around for exports and adding a bootstrap phase to adapters
